+++ src/dotnet/APIView/APIViewWeb/Controllers/UserProfileController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using APIViewWeb.Models;
 using System.Threading.Tasks;
+using System.Collections.Generic;
 
 namespace APIViewWeb.Controllers
 {
@@ -14,16 +15,18 @@ namespace APIViewWeb.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult> Update(string email)
+        public async Task<ActionResult> Update(string email, string[] languages)
         {
             UserProfileModel profile = await _userProfileManager.tryGetUserProfileAsync(User);
 
+            HashSet<string> Languages = new HashSet<string>(languages);
             if(profile.UserName == null)
             {
-                await _userProfileManager.createUserProfileAsync(User, email);
+                await _userProfileManager.createUserProfileAsync(User, email, Languages);
             } else
             {
                 await _userProfileManager.updateEmailAsync(User, email);
+                await _userProfileManager.updateLanguagesAsync(User, Languages);
             }
             return Ok();
         }
