+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Index.cshtml
@@ -12,7 +12,6 @@
 @section Scripts
 {
     <script type="text/javascript" src="@Url.Content("~/reviews.js")" asp-append-version="true"></script>
-    <script type="text/javascript" src="@Url.Content("~/userProfile.js")" asp-append-version="true"></script>
     <script type="text/javascript" src="@Url.Content("~/fileInput.js")" asp-append-version="true"></script>
 }
 <nav class="navbar-light shadow-sm bg-light sticky-top">
