+++ src/dotnet/APIView/APIViewWeb/Pages/Shared/_Layout.cshtml
@@ -72,28 +72,28 @@
                         }
                     </ul>
                     <ul class="navbar-nav ml-auto">
-                        @if (User.Identity.IsAuthenticated) {
-                            <li class="nav-item mt-1">
-                                <select class="selectpicker show-tick p-0" data-style="btn-light btn-sm border m-0" data-width="fit" id="theme-selector" data-tick-icon="fa-solid fa-check">
-                                    @foreach(var item in themes)
+                        <li class="nav-item mt-1">
+                            <select class="selectpicker show-tick p-0" data-style="btn-light btn-sm border m-0" data-width="fit" id="theme-selector" data-tick-icon="fa-solid fa-check">
+                                @foreach(var item in themes)
+                                {
+                                    var displayName = item.Replace("-theme", "");
+                                    if (item.Equals(theme))
                                     {
-                                        var displayName = item.Replace("-theme", "");
-                                        if (item.Equals(theme))
-                                        {
-                                            <option value="@item" selected>@displayName</option>
-                                        }
-                                        else
-                                        {
-                                            <option value="@item">@displayName</option>
-                                        }
+                                        <option value="@item" selected>@displayName</option>
                                     }
-                                </select>
+                                    else
+                                    {
+                                        <option value="@item">@displayName</option>
+                                    }
+                                }
+                            </select>
+                        </li>
+                        @if (User.Identity.IsAuthenticated) {
+                            <li class="nav-item">
+                                <a asp-page="./Profile" class="nav-link">Profile</a>
                             </li>
                             <li class="nav-item">
                                 <a asp-action="Logout" asp-controller="Account" class="nav-link">Log Out</a>
-                            </li>
-                            <li class="nav-item">
-                                <a  asp-page="./Profile" class="nav-link">Profile</a>
                             </li>
                         }
                         <li class="nav-item">
@@ -149,7 +149,6 @@
                 integrity="sha384-xrRywqdh3PHs8keKZN+8zzc5TX0GRTLCcmivcbNJWm2rs5C8PRhcEn3czEjhAO9o">
         </script>
     </environment>
-    <script src="~/navbar.js" asp-append-version="true"></script>
     <!-- Boostrap-Select-->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.14/dist/js/bootstrap-select.min.js"
             crossorigin="anonymous">
