+++ src/dotnet/APIView/APIViewWeb/Pages/Shared/_UserProfileFormPartial.cshtml
@@ -1,16 +1,54 @@
 ﻿@using APIViewWeb.Models;
 @model UserProfileModel;
-
-@if(Model != null)
-{
-    <form data-post-update="userProfile" asp-controller="UserProfile" asp-action="Update" method="post">
-        <div class="form-group">
-            <label for="UserName" class="text-muted"> GitHub User Profile </label>
-            <input name="UserName" class="form-control" type="text" readonly value="@Model.UserName" /><br />
-            <label for="Email" class="text-muted"> Email </label>
-            <input name="Email" class="form-control" type="email" value="@Model.Email" /><br />
-            <hr />
-            <input class="btn btn-primary" type="submit" value="Save Changes" />
-        </div>
-    </form>
+@{
+    string[] Languages = { "c", "c#", "c++", "go", "java", "javascript", "python", "protocol", "swift" };
+    var userPreference = (TempData["UserPreference"] as UserPreferenceModel) ?? new UserPreferenceModel();
+    List<string> themes = new List<string>()
+    {
+        "light-theme",
+        "dark-theme",
+        "dark-solarized-theme"
+    };
+    var theme = userPreference.Theme ?? "light-theme";
 }
+
+<form data-post-update="userProfile" asp-controller="UserProfile" asp-action="Update" method="post">
+    <div class="d-flex flex-column align-items-center">
+        <img username="@User.GetGitHubLogin()" class="w-25 rounded-circle my-3" style="min-width:80px; max-width:160px" aria-label="GitHub User Avatar" />
+        <h4 class="w-100 text-center">@Model.UserName</h4>
+    </div>
+    <hr />
+    <div class="form-group form-row">
+        <label for="Email" class="col-sm-2 font-weight-bold" style="line-height:48px"> Email </label>
+        <input name="Email" class="form-control form-control-lg col-sm-10" type="email" value="@Model.Email" required /><br />
+    </div>
+    <div class="form-group form-row">
+        <label for="Languages" class="col-sm-2 font-weight-bold" style="line-height: 38px"> Review Languages </label>
+        <select name="Languages" class="selectpicker col-sm-10 px-0" multiple>
+            @foreach(string language in Languages) {
+                <option value="@language" selected="@(Model.Languages.Contains(language) ? true : false)">@language</option>
+            }
+        </select>
+    </div>
+    <div class="form-group form-row">
+        <label for="Theme" class="col-sm-2 font-weight-bold" style="line-height: 38px"> Theme </label>
+        <select name="Theme" class="selectpicker show-tick col-sm-10 px-0" id="theme-selector" data-tick-icon="fa-solid fa-check">
+            @foreach(var item in themes)
+            {
+                var displayName = item.Replace("-theme", "");
+                if (item.Equals(theme))
+                {
+                    <option value="@item" selected>@displayName</option>
+                }
+                else
+                {
+                    <option value="@item">@displayName</option>
+                }
+            }
+        </select>
+    </div>
+    <hr />
+    <div class="d-flex justify-content-end align-items-center">
+        <input class="btn btn-primary" type="submit" value="Save Changes" />
+    </div>
+</form>
